feat: move agent into core
