+++ packages/core/src/global/index.ts
@@ -9,4 +9,4 @@ export type {
   LLMToolResultEvent,
   LlamaIndexEventMaps,
 } from "./settings/callback-manager";
-export type { JSONArray, JSONObject, JSONValue } from "./type";
+export type { JSONArray, JSONObject, JSONValue, UUID } from "./type";
