+++ packages/core/tsconfig.json
@@ -8,7 +8,7 @@
     "moduleResolution": "Bundler",
     "skipLibCheck": true,
     "strict": true,
-    "lib": ["ESNext", "DOM"],
+    "lib": ["ESNext", "DOM", "DOM.AsyncIterable"],
     "types": ["node"]
   },
   "include": ["./src"],
