+++ packages/llamaindex/src/agent/index.ts
@@ -1,10 +1,9 @@
+export * from "@llamaindex/core/agent";
 export {
   AnthropicAgent,
   AnthropicAgentWorker,
   type AnthropicAgentParams,
 } from "./anthropic.js";
-export { AgentRunner, AgentWorker, type AgentParamsBase } from "./base.js";
-export { LLMAgent, LLMAgentWorker, type LLMAgentParams } from "./llm.js";
 export {
   OpenAIAgent,
   OpenAIAgentWorker,
@@ -15,8 +14,6 @@ export {
   ReActAgent,
   type ReACTAgentParams,
 } from "./react.js";
-export { type TaskHandler } from "./types.js";
-export { callTool, stepTools, stepToolsStreaming } from "./utils.js";
 
 // todo: ParallelAgent
 // todo: CustomAgent
