+++ packages/llamaindex/src/agent/openai.ts
@@ -1,6 +1,10 @@
+import {
+  LLMAgent,
+  LLMAgentWorker,
+  type LLMAgentParams,
+} from "@llamaindex/core/agent";
 import { OpenAI } from "@llamaindex/openai";
 import { Settings } from "../Settings.js";
-import { LLMAgent, LLMAgentWorker, type LLMAgentParams } from "./llm.js";
 
 // This is likely not necessary anymore but leaving it here just incase it's in use elsewhere
